+++ Classes/Service/OneTimeAccountConnector.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Service;
 
+use Psr\Http\Message\ServerRequestInterface;
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Frontend\Authentication\FrontendUserAuthentication;
-use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
 /**
  * Connects to FE user accounts and sessions data created by the "onetimeaccount" extension.
@@ -18,20 +18,10 @@ class OneTimeAccountConnector implements SingletonInterface
      */
     private $frontEndUserAuthentication;
 
-    /**
-     * @throws \RuntimeException
-     */
     public function __construct()
     {
-        $frontEndController = $GLOBALS['TSFE'] ?? null;
-        if (!$frontEndController instanceof TypoScriptFrontendController) {
-            throw new \RuntimeException('No frontend found.', 1668702167);
-        }
-        $frontEndUserAuthentication = $frontEndController->fe_user;
-        if (!$frontEndUserAuthentication instanceof FrontendUserAuthentication) {
-            throw new \RuntimeException('Frontend found, but without a FE user authentication.', 1668702517);
-        }
-
+        $frontEndUserAuthentication = $this->getRequest()->getAttribute('frontend.user');
+        \assert($frontEndUserAuthentication instanceof FrontendUserAuthentication);
         $this->frontEndUserAuthentication = $frontEndUserAuthentication;
     }
 
@@ -60,5 +50,13 @@ class OneTimeAccountConnector implements SingletonInterface
         if (\is_int($this->getOneTimeAccountUserUid())) {
             $this->frontEndUserAuthentication->setAndSaveSessionData('onetimeaccountUserUid', null);
         }
+    }
+
+    private function getRequest(): ServerRequestInterface
+    {
+        $request = $GLOBALS['TYPO3_REQUEST'];
+        \assert($request instanceof ServerRequestInterface);
+
+        return $request;
     }
 }